We were using a short to store an int and if the upnp client was using a port too high it was causing a failure for us to respond to subscribe events.
git-svn-id: http://svn.mythtv.org/svn/trunk@17527 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythupnp/eventing.cpp
@@ -314,7 +314,7 @@ void Eventing::NotifySubscriber( SubscriberInfo *pInfo )
         // Build Message Header 
         // ----------------------------------------------------------------------
 
-        short   nPort = pInfo->qURL.hasPort() ? pInfo->qURL.port() : 80;
+        int   nPort = pInfo->qURL.hasPort() ? pInfo->qURL.port() : 80;
         QString sHost = QString( "%1:%2" ).arg( pInfo->qURL.host() )
                                           .arg( nPort );
